extract plotting
